carved out models module and added some initial classes for observability
