+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/execution/StageChainRegistry.java
@@ -13,10 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.stage;
+package com.github.nishgpt.chainexecutor.core.execution;
 
-import com.github.nishgpt.chainexecutor.core.exceptions.ChainExecutorException;
-import com.github.nishgpt.chainexecutor.core.exceptions.ErrorCode;
+import com.github.nishgpt.chainexecutor.models.error.ChainExecutorException;
+import com.github.nishgpt.chainexecutor.models.error.ErrorCode;
+import com.github.nishgpt.chainexecutor.models.stage.Stage;
+import com.github.nishgpt.chainexecutor.models.stage.StageChain;
+import com.github.nishgpt.chainexecutor.models.stage.StageChainIdentifier;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
