+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/execution/StageExecutorFactory.java
@@ -13,8 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.execution;
+package com.github.nishgpt.chainexecutor.core.execution;
 
+import com.github.nishgpt.chainexecutor.models.execution.StageExecutorKey;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
 import com.google.inject.Injector;
