+++ chain-executor-core/src/main/java/com/github/nishgpt/chainexecutor/core/observability/MethodCriticality.java
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.stage;
+package com.github.nishgpt.chainexecutor.core.observability;
 
-public interface StageChainIdentifier {
-
+public enum MethodCriticality {
+  HIGH,
+  MEDIUM,
+  LOW,
+  ;
 }
