+++ chain-executor-models/pom.xml
@@ -26,7 +26,7 @@
     <version>0.1.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>chain-executor-observability</artifactId>
+  <artifactId>chain-executor-models</artifactId>
 
   <properties>
     <maven.compiler.source>17</maven.compiler.source>
@@ -34,4 +34,10 @@
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
   </properties>
 
+  <dependencies>
+    <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+    </dependency>
+  </dependencies>
 </project>
