+++ chain-executor-models/src/main/java/com/github/nishgpt/chainexecutor/models/error/ErrorCode.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright(c) 2023 Nishant Gupta (nishant141077@gmail.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.nishgpt.chainexecutor.models.error;
+
+import lombok.Getter;
+
+@Getter
+public enum ErrorCode {
+  INVALID_CHAIN,
+  EXECUTION_ERROR,
+  INVALID_EXECUTION_STAGE,
+  EXECUTION_NOT_COMPLETED,
+  PRE_EXECUTION_FAILED
+}
