+++ chain-executor-models/src/main/java/com/github/nishgpt/chainexecutor/models/execution/ExecutorAuxiliaryKey.java
@@ -13,11 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.execution;
+package com.github.nishgpt.chainexecutor.models.execution;
 
-import com.github.nishgpt.chainexecutor.core.models.stage.Stage;
+import java.io.Serializable;
 
-public interface StageExecutionRequest {
+/**
+ * Custom Key/Identifier interface for stage executor. This in combination with Stage is to be used for registering and
+ * fetching StageExecutor
+ */
+public interface ExecutorAuxiliaryKey extends Serializable {
 
-  Stage getStage();
 }
