+++ chain-executor-models/src/main/java/com/github/nishgpt/chainexecutor/models/stage/StageChainIdentifier.java
@@ -13,14 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.execution;
+package com.github.nishgpt.chainexecutor.models.stage;
 
-import java.io.Serializable;
-
-/**
- * Custom Key/Identifier interface for stage executor. This in combination with Stage is to be used for registering and
- * fetching StageExecutor
- */
-public interface ExecutorAuxiliaryKey extends Serializable {
+public interface StageChainIdentifier {
 
 }
