+++ chain-executor-models/src/main/java/com/github/nishgpt/chainexecutor/models/stage/StageStatus.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.nishgpt.chainexecutor.core.models.stage;
+package com.github.nishgpt.chainexecutor.models.stage;
 
 import java.util.Set;
 import lombok.Getter;
